# packages/shared/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/companies:
    post:
      operationId: CompaniesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyDto"
      responses:
        "201":
          description: ""
      security: &a4
        - bearer: []
      summary: Create a company and attach it to my account
      tags: &a5
        - companies
  /api/v1/companies/mine:
    get:
      operationId: CompaniesController_mine
      parameters: []
      responses:
        "200":
          description: ""
      security: *a4
      summary: My company with addresses
      tags: *a5
  /api/v1/companies/mine/addresses:
    post:
      operationId: CompaniesController_addAddress
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressDto"
      responses:
        "201":
          description: ""
      security: *a4
      summary: Add an address to my company
      tags: *a5
  /api/v1/companies/mine/addresses/{id}:
    patch:
      operationId: CompaniesController_updateAddress
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressDto"
      responses:
        "200":
          description: ""
      security: *a4
      summary: Update one of my company addresses
      tags: *a5
    delete:
      operationId: CompaniesController_deleteAddress
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a4
      summary: Soft-delete one of my company addresses
      tags: *a5
...
  /api/v1/documents/upload:
    post:
      operationId: DocumentsController_upload
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                kind:
                  type: string
                  enum:
                    - rfq_attachment
                    - quote_attachment
                    - order_document
                    - shipment_document
                    - compliance
                    - other
                ownerType:
                  type: string
                  enum:
                    - rfq
                    - order
                    - shipment
                    - support_ticket
                ownerId:
                  type: string
                  format: uuid
              required:
                - file
                - ownerType
                - ownerId
      responses:
        "201":
          description: ""
      security: &a21
        - bearer: []
      summary: Upload a document (10 MB max; stored privately, random filename)
      tags: &a22
        - documents
  /api/v1/documents/{id}/download-url:
    get:
      operationId: DocumentsController_downloadUrl
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a21
      summary: Mint a 15-minute signed download URL (authorized callers only)
      tags: *a22
  /api/v1/documents/download:
    get:
      operationId: DocumentsController_download
      parameters:
        - name: token
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a21
      summary: Download via signed temporary URL (15-min expiry, HMAC-verified)
      tags: *a22
...
info:
  title: EGT B2B API
  description: Eagle Goods Trading Co. — sourcing & export backend API (v1)
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateCompanyDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - buyer
            - supplier
            - both
          default: buyer
        country:
          type: string
        city:
          type: string
        website:
          type: string
      required:
        - name
    CreateAddressDto:
      type: object
      properties:
        label:
          type: string
          default: Head office
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
        isDefault:
          type: boolean
          default: false
      required:
        - line1
        - city
        - country
...